+++ bigtable-dataflow-parent/bigtable-hbase-beam/src/main/java/com/google/cloud/bigtable/beam/CloudBigtableIO.java
@@ -38,7 +38,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.NoSuchElementException;
 import java.util.concurrent.atomic.AtomicLong;
-import org.apache.beam.sdk.annotations.Experimental;
 import org.apache.beam.sdk.coders.CannotProvideCoderException;
 import org.apache.beam.sdk.coders.Coder;
 import org.apache.beam.sdk.coders.CoderRegistry;
@@ -130,7 +129,6 @@ import org.slf4j.LoggerFactory;
  *          .build()));
  * }</pre>
  */
-@Experimental
 public class CloudBigtableIO {
 
   /**
@@ -1003,15 +1001,18 @@ public class CloudBigtableIO {
 
     @FinishBundle
     public void finishBundle(FinishBundleContext c) throws Exception {
-      for (BufferedMutator bufferedMutator : mutators.values()) {
-        try {
-          bufferedMutator.flush();
-        } catch (RetriesExhaustedWithDetailsException exception) {
-          logExceptions(c, exception);
-          rethrowException(exception);
+      try {
+        for (BufferedMutator bufferedMutator : mutators.values()) {
+          try {
+            bufferedMutator.close();
+          } catch (RetriesExhaustedWithDetailsException exception) {
+            logExceptions(c, exception);
+            rethrowException(exception);
+          }
         }
+      } finally {
+        mutators.clear();
       }
-      mutators.clear();
     }
   }